Refactor event handling to use repository pattern
This commit introduces significant refactoring of the codebase, focusing on restructuring event-related classes and filters. The `Xpandables.Net.Events.Entities` namespace has been removed in favor of `Xpandables.Net.Repositories.Filters`, promoting a repository-oriented architecture.

Key changes include:
- Modifications to `EventConverter` classes to align with the new repository structure.
- Updates to `EventPublisherSubscriber` and `EventScheduler` for improved filtering mechanisms.
- Upgraded dependencies in the project file for compatibility and new features.
- Refinements to `IEventFilter` and `IEntityFilter` interfaces to support asynchronous operations and enhanced filtering capabilities.

These changes aim to improve maintainability, scalability, and testability by decoupling event entities from their handling logic and introducing a more robust filtering system.
