+++ Xpandables.Net.Api/Accounts/Endpoints/GetOperationsAccount/GetOperationsAccountQueryHandler.cs
@@ -3,7 +3,7 @@
 using Xpandables.Net.Api.Accounts.Events;
 using Xpandables.Net.Commands;
 using Xpandables.Net.Events;
-using Xpandables.Net.Events.Filters;
+using Xpandables.Net.Repositories.Filters;
 
 namespace Xpandables.Net.Api.Accounts.Endpoints.GetOperationsAccount;
