+++ Xpandables.Net.Api/Accounts/Persistence/DataContextEventSqlServerBuilder.cs
@@ -4,7 +4,6 @@ using Microsoft.EntityFrameworkCore.Metadata.Conventions;
 
 using Xpandables.Net.DependencyInjection;
 using Xpandables.Net.Events.Converters;
-using Xpandables.Net.Events.Entities;
 using Xpandables.Net.Repositories;
 
 namespace Xpandables.Net.Api.Accounts.Persistence;
