+++ Xpandables.Net.EntityFramework/Repositories/DataContextEvent.cs
@@ -18,7 +18,6 @@
 using Microsoft.EntityFrameworkCore;
 
 using Xpandables.Net.Events;
-using Xpandables.Net.Events.Entities;
 
 namespace Xpandables.Net.Repositories;
 /// <summary>
