+++ Xpandables.Net.Test/UnitTests/EventPublisherSubscriberUnitTest.cs
@@ -5,7 +5,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Xpandables.Net.Commands;
 using Xpandables.Net.DependencyInjection;
 using Xpandables.Net.Events;
-using Xpandables.Net.Events.Defaults;
 using Xpandables.Net.Operations;
 
 namespace Xpandables.Net.Test.UnitTests;
